testing paco -s

+++ ft_printf.h
@@ -8,7 +8,7 @@
 int		ft_printf(const char *, ...);
 void	ft_putchar(char c, int *printed);
 void	ft_putstr(char *s, int *printed);
-void	ft_puthex(unsigned long long num, char c, int *printed);
+void	ft_puthex(size_t num, char c, int *printed);
 void	ft_putnbr(int n, int *printed);
 void	ft_putunsignednbr(unsigned int n, int *printed);
